requirements.txt
langchain
wikipedia
langchain-community
langchain-openai    
langchain-anthropic
pyhton-dotenv
pydantic
ddgs
fastapi